+++ lnd_methods/peers/index.js
@@ -2,5 +2,14 @@ const addPeer = require('./add_peer');
 const getPeers = require('./get_peers');
 const removePeer = require('./remove_peer');
 const subscribeToPeers = require('./subscribe_to_peers');
+const updateAlias = require('./update_alias');
+const updateColor = require('./update_color');
 
-module.exports = {addPeer, getPeers, removePeer, subscribeToPeers};
+module.exports = {
+  addPeer,
+  getPeers,
+  removePeer,
+  subscribeToPeers,
+  updateAlias,
+  updateColor,
+};